Machina.FFXIV v2.3.0.3: Updates to support Global patch 5.58 and Korean patch 5.45.  Fix to Waymark packet size.

+++ Machina.FFXIV/Headers/Korean/WaymarkType.cs
@@ -18,7 +18,7 @@ using System;
 namespace Machina.FFXIV.Headers.Korean
 {
     [Flags]
-    public enum WaymarkType : uint
+    public enum WaymarkType : byte
     {
         A = 0x1,
         B = 0x2,
